chore: Renames main file to be more standard.
